pyproject.toml
[project]
name = "fata-cognita"
version = "0.1.0"
description = "Neural Actuarial Trajectory Model — VAE+Transformer for longitudinal life-trajectory modeling"
requires-python = ">=3.14"
dependencies = [
    "torch>=2.5",
    "numpy>=2.0",
    "pandas>=2.2",
    "scikit-learn>=1.5",
    "pyarrow>=17.0",
    "pyyaml>=6.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.28",
    "streamlit>=1.40",
    "plotly>=5.24",
    "matplotlib>=3.9",
    "umap-learn>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fata_cognita"]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["fata_cognita"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
